requirements.txt
Django>=5.2
python-dotenv>=1.0
google-generativeai>=0.8